[struct siginfo] (si_band): Change type to `long int'.

+++ sysdeps/unix/sysv/linux/bits/siginfo.h
@@ -93,7 +93,7 @@ typedef struct siginfo
 	/* SIGPOLL.  */
 	struct
 	  {
-	    int si_band;	/* Band event for SIGPOLL.  */
+	    long int si_band;	/* Band event for SIGPOLL.  */
 	    int si_fd;
 	  } _sigpoll;
       } _sifields;
